nigeria-nanny

+++ shop/new-password.php
@@ -12,29 +12,32 @@ if(!Input::exists('get') && Input::get('tid'))
 
 if(Input::post('reset_password'))
 {
-    $validate = new DB(); 
-    $validation = $validate->validate([
-        'new_password' => 'required|min:6|max:12',
-        'confirm_password' => 'required|min:6|max:12|match:new_password'
-    ]);
-
-    if(!$validation->passed())
+    if(Token::check())
     {
-        return back();
-    }
+        $validate = new DB(); 
+        $validation = $validate->validate([
+            'new_password' => 'required|min:6|max:12',
+            'confirm_password' => 'required|min:6|max:12|match:new_password'
+        ]);
 
-    $reset = $connection->select('reset_password')->where('reset_token', Input::get('reset_token'))->first();
-    if($reset)
-    {
-        $update_passowrd = $connection->update('users', [
-            'password' => password_hash(Input::get('new_password'), PASSWORD_DEFAULT)
-        ])->where('email', $reset->reset_email)->save();
-
-        if($update_passowrd)
+        if(!$validation->passed())
         {
-            $connection->delete('reset_password')->where('reset_token', Input::get('reset_token'))->save();
-            Session::delete('get_passsword');
-            return Redirect::to('login.php', ['success', 'Password reset successfully, you can login!']);
+            return back();
+        }
+
+        $reset = $connection->select('reset_password')->where('reset_token', Input::get('reset_token'))->first();
+        if($reset)
+        {
+            $update_passowrd = $connection->update('users', [
+                'password' => password_hash(Input::get('new_password'), PASSWORD_DEFAULT)
+            ])->where('email', $reset->reset_email)->save();
+
+            if($update_passowrd)
+            {
+                $connection->delete('reset_password')->where('reset_token', Input::get('reset_token'))->save();
+                Session::delete('get_passsword');
+                return Redirect::to('login.php', ['success', 'Password reset successfully, you can login!']);
+            }
         }
     }
 }
@@ -95,6 +98,7 @@ if(Input::post('reset_password'))
                 </div>
                 <button type="submit" name="reset_password" class="btn btn-log btn-block btn-thm2">Reset password</button>
                 <hr>
+                <?= csrf_token() ?>
             </form>
         </div>
     </div>
